🐛 Fix `context.range` updates
The `Selection` class has a (hidden) feature that allows blots to update
the selection from their `optimize()` hooks by setting
[`context.range`][1].

This is currently used by [`Embed`s][2] for example, to correct the
selection if their guards are edited.

The current implementation has a couple of bugs, though.

The first bug is that the selection is currently during the
`SCROLL_OPTIMIZE` event. However, since this is called *before* the
`SCROLL_UPDATE` event, the `Selection` then clobbers any changes made by
trying to [put back the previous selection][3].

The second bug is more subtle, and I sadly couldn't write a failing test
case for it, since it seems Android-specific: if a blot removes itself
in response to user input, *and* also updates `context.range`, then the
browser will attempt to update the selection in response to the user
input, which happens after the Quill events are done processing.

In order to update the selection after this, we move the `context.range`
update into a `setTimeout()` call, which should let the browser finish
processing the current user input before we try to update the selection.

Using a `setTimeout()` isn't ideal, since it introduces some
non-deterministic behaviour, but it looks like this workaround was also
added to deal with [`compositionend`][4].

[1]: https://github.com/reedsy/quill/blob/d2ea83a81fa6104bbf9be4349a73a74b97f33c26/core/selection.ts#L91
[2]: https://github.com/reedsy/quill/blob/d2ea83a81fa6104bbf9be4349a73a74b97f33c26/blots/embed.ts#L79
[3]: https://github.com/reedsy/quill/blob/d2ea83a81fa6104bbf9be4349a73a74b97f33c26/core/selection.ts#L60-L89
[4]: https://github.com/quilljs/quill/commit/e7809196f851794a53f12033cedf73054b02e8d8