💥 Remove `Emitter.listeners` type clash
At the moment, since `@ts-expect-error` [doesn't apply][1] to type
definitions, we get a downstream compilation error:

```
TS2416: Property 'listeners' in type 'Emitter' is not assignable to the same property in base type 'EventEmitter<string, any>'.
  Type 'Record<string, { node: Node; handler: Function; }[]>' is not assignable to type '<T extends string>(event: T) => ((...args: any[]) => void)[]'.
    Type 'Record<string, { node: Node; handler: Function; }[]>' provides no match for the signature '<T extends string>(event: T): ((...args: any[]) => void)[]'.

19     listeners: Record<string, {
```

This change removes the clash by renaming the internal variable. Note
that, since `listeners` is `public`, this is technically breaking.

[1]: https://github.com/microsoft/TypeScript/issues/38628